Move Deref to module.
